Write a function that prints an integer.

+++ 0x06-pointers_arrays_strings/101-print_number.c
@@ -8,7 +8,7 @@
 
 void print_number(int n)
 {
-	unsigned int div = 1;
+	long long div = 1;
 
 	if (n < 0)
 	{
